rename to terminate-unhealthy-nodes
