+++ service/controller/resource/terminateunhealthynode/delete.go
@@ -1,4 +1,4 @@
-package nodeautorepair
+package terminateunhealthynode
 
 import (
 	"context"
